Move filtering OS accroding to architecture to separate function and move VM architecture discover to createOsList function
